+++ src/config.rs
@@ -56,8 +56,9 @@ use crate::config_http::{
 };
 pub use crate::config_load_balance::{
     LoadBalanceConfig, LoadBalanceHealthCheckConfig, LoadBalanceHealthCheckExpectedHeader,
-    LoadBalanceHealthCheckProtocol, LoadBalancePassiveHealthConfig, LoadBalanceRetryConfig,
-    LoadBalanceSelection, LoadBalanceSlowStartConfig,
+    LoadBalanceHealthCheckExpectedStatusRange, LoadBalanceHealthCheckProtocol,
+    LoadBalancePassiveHealthConfig, LoadBalanceRetryConfig, LoadBalanceSelection,
+    LoadBalanceSlowStartConfig,
 };
 pub use crate::config_loader::ConfigLoadError;
 #[cfg(feature = "load-balancer")]